load_betterbiblatex_bib {citr}	R Documentation

Load bibliography from Zotero

Description

Load bibliography from Zotero

Usage

load_betterbiblatex_bib(
  encoding,
  betterbiblatex_format = "bibtex",
  exclude_betterbiblatex_library = NULL,
  increment_progress = FALSE
)

Arguments

encoding
Character. Encoding of the Bib(La)TeX-file. See the connection.

betterbiblatex_format
Character. Bibliography format to export from Zotero/Juris-M. Can be either "bibtex" or "biblatex". Ignored if use_betterbiblatex = FALSE. Requires that the Better BibTeX is installed and Zotero/Juris-M is running.

exclude_betterbiblatex_library
Character. A vector of Zotero/Juris-M library names not to query.

increment_progress
logical switch which will use the shiny function incProgress when TRUE. FALSE by default.

Details

This function loads into RAM the bibliography stored in Zotero. May take a several seconds if you have many hundreds of references.

Examples

## Not run: 
b <- load_betterbiblatex_bib(encoding = "UTF-8")

## End(Not run)

[Package citr version 0.3.2 Index]