md_cite_zotero {citr}	R Documentation

Insert citation in Markdown format from Zotero

Description

Look up entries in Zotero database and insert citation in Markdown format.

Usage

md_cite_zotero(in_paren = TRUE)

Arguments

in_paren
Logical. Determines if citation is in parentheses.

Details

The path to the Bib(La)TeX-file can be set in the global options and is set to references.bib when the package is loaded. Once the path is changed in the RStudio addin, the global option is updated.

Value

If the bibliography contains exactly one match the formatted citation is returned, otherwise returns NULL. md_cite returns an in-text citation ("@foo2016"), md_cite returns an in-parenthesis citation ("[@foo2016]").

See Also

insert_citation

Examples

## Not run: 
  md_cite_zotero()

## End(Not run)

[Package citr version 0.3.2 Index]